refactor: Extract gbm framebuffer code
